requirements.txt
fastapi
uvicorn
snowflake-connector-python
pandas
numpy
pydantic
python-dotenv
scikit-learn
statsmodels
